Move App, BookForm and BookList to containers fold
