+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import rootReducer from './reducers';
-import App from './components/App';
+import App from './containers/App';
 
 const initialState = {
   bookReducer: [
